# docs/openapi.yml
openapi: 3.0.0
info:
  title: Song Rating API
  description: API for the song ratings app.
  version: 1.0.0
servers:
  - url: "https://songrat.ing/api"
paths:
...
  /game/check_code/{code}:
    get:
      summary: Get whether a code is valid.
      description: Get whether an invite, player, or admin code is valid.
      operationId: checkCode
      parameters:
        - name: code
          description: An invite, player, or admin code.
          in: path
          required: true
          schema:
            type: string
            pattern: "^[IPA][a-zA-Z0-9]{15}$"
      responses:
        "200":
          description: Code check succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_valid:
                    description: Whether the code is valid.
                    type: boolean
        "400":
          description: Invalid input.
        "500":
          description: Internal server error.
  /game/new:
    post:
      summary: Create a new game.
      description: Create a new game by supplying game info.
      operationId: createGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                game_name:
                  description: The name for this game.
                  type: string
                  minLength: 1
                  maxLength: 255
                game_description:
                  description: The description for this game (can be empty).
                  type: string
                  maxLength: 2500
                min_songs_per_playlist:
                  description:
                    The minimum number of songs allowed in any playlist. Must be <=
                    'max_songs_per_playlist' if 'max_songs_per_playlist' isn't null.
                  type: integer
                  minimum: 1
                  maximum: 100
                max_songs_per_playlist:
                  description:
                    The maximum number of songs allowed in any playlist. Must be >=
                    'min_songs_per_playlist' if 'min_songs_per_playlist' isn't null.
                  type: integer
                  minimum: 1
                  maximum: 100
                require_playlist_link:
                  description: Whether players are required to provide a playlist link.
                  type: boolean
              required:
                - game_name
                - require_playlist_link
      responses:
        "201":
          description: Successfully created game.
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin_code:
                    description: The admin code for managing this game.
                    type: string
                    pattern: "^A[a-zA-Z0-9]{15}$"
                  invite_code:
                    description: The code players can use to join this game.
                    type: string
                    pattern: "^I[a-zA-Z0-9]{15}$"
        "400":
          description: Invalid input.
        "500":
          description: Internal server error.
  /game/peek/{invite_code}:
    get:
      summary: Get some information about a game.
      description: Retrieve some information about a game with an invite code.
      operationId: peekGameInfo
      parameters:
        - name: invite_code
          description: An invite code for the game to get info about.
          in: path
          required: true
          schema:
            type: string
            pattern: "^I[a-zA-Z0-9]{15}$"
      responses:
        "200":
          description: Successfully retrieved game information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  game_name:
                    description: The name of this game.
                    type: string
                    minLength: 1
                    maxLength: 255
                  game_description:
                    description: The description of this game.
                    type: string
                  min_songs_per_playlist:
                    description: The minimum number of songs allowed in any playlist.
                    type: integer
                    minimum: 1
                    maximum: 100
                  max_songs_per_playlist:
                    description: The maximum number of songs allowed in any playlist.
                    type: integer
                    minimum: 1
                    maximum: 100
                  require_playlist_link:
                    description: Whether players are required to provide a playlist link.
                    type: boolean
        "400":
          description: Invalid input.
        "404":
          description: Unknown 'invite_code' supplied.
        "500":
          description: Internal server error.
  /game/join/{invite_code}:
    post:
      summary: Create a new player by joining a game.
      description: Create a new player by joining a game with a game's invite code.
      operationId: joinGame
      parameters:
        - name: invite_code
          description: An invite code for the game to create a new player for.
          in: path
          required: true
          schema:
            type: string
            pattern: "^I[a-zA-Z0-9]{15}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_name:
                  description: The name for this player.
                  type: string
                  minLength: 1
                  maxLength: 255
                playlist_link:
                  description: The link for this player's playlist.
                  nullable: true
                  type: string
                  minLength: 1
                  maxLength: 255
                songs:
                  description:
                    A list of all songs in this player's playlist (must be within range of
                    'min_songs_per_playlist' and 'max_songs_per_playlist').
                  type: array
                  items:
                    description: A song.
                    type: object
                    properties:
                      title:
                        description: The name of this song.
                        type: string
                        minLength: 1
                        maxLength: 255
                      artist:
                        description: The artist of this song.
                        type: string
                        minLength: 1
                        maxLength: 255
                    required:
                      - title
                      - artist
                  minItems: 1
                  maxItems: 100
              required:
                - player_name
                - songs
      responses:
        "201":
          description: Game joined, player created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  player_code:
                    description: The player code for managing this player.
                    type: string
        "400":
          description: Invalid input.
        "404":
          description: Unknown 'invite_code' supplied.
        "409":
          description:
            Invalid number of songs supplied (must be within range of 'min_songs_per_playlist' and
            'max_songs_per_playlist'), name is already taken, playlist link required but not
            provided, or the max player count has been reached/exceeded.
        "410":
          description: The 'invite_code' is for a game that has already begun.
        "500":
          description: Internal server error.

  # /admin
...